# models/user.js
const mongoose = require("mongoose");

// user model schema
const userSchema = mongoose.Schema({
  name: {
    type: String,
    required: [true, "Please provide a name"],
    minLength: 3,
    maxLength: 50,
  },
  username: {
    type: String,
    required: [true, "Please provide an username"],
    match: [
      /^(?=[a-zA-Z0-9._]{8,20}$)(?!.*[_.]{2})[^_.].*[^_.]$/,
      "Please provide a valid username",
    ],
    unique: true,
  },
  password: {
    type: String,
    required: [true, "Please provide a password"],
    minLength: 6,
  },
});

module.exports = mongoose.model("User", userSchema);